fix: move match utility functions to separate file to resolve ESLint errors
The getMatchWinner and getMatchMargin functions were exported from
Fixtures.tsx alongside the Fixtures component. This caused two
react-refresh/only-export-components ESLint errors since fast-refresh
requires component files to only export components.

Fix: extract MatchWinner type, getMatchWinner and getMatchMargin into
src/utils/matchUtils.ts, import them into Fixtures.tsx for internal use,
and update the test file to import from the new location.

108 tests pass; lint clean.
